Implement A-Packages
* Organize classes into packages
