+++ src/main/java/duke/storage/Storage.java
@@ -1,6 +1,19 @@
-import java.io.*;
-import java.util.ArrayList;
-import java.util.List;
+package duke.storage;
+
+import duke.exception.DukeException;
+import duke.task.Deadline;
+import duke.task.Event;
+import duke.task.Task;
+import duke.task.TaskList;
+import duke.task.TaskType;
+import duke.task.ToDo;
+
+import java.io.BufferedReader;
+import java.io.BufferedWriter;
+import java.io.File;
+import java.io.FileReader;
+import java.io.FileWriter;
+import java.io.IOException;
 
 /**
  * A class that handles create, read and write file operations.
